scripts/requirements.txt
    ipykernel==6.29.5
    langchain==0.3.25
    langchain-experimental==0.3.4
    langchain-openai==0.3.23
    langgraph==0.4.8
    pandas==2.3.0
    streamlit==1.45.1
    tabulate==0.9.0
    dotenv ==0.9.9
    google-api-python-client==2.172.0
    google-auth==2.40.3
    google-auth-oauthlib==1.2.2
    langchain-neo4j==0.4.0
